// lib/game/types/enums.ts
export enum Direction {
    UP = 0,
    RIGHT = 1,
    DOWN = 2,
    LEFT = 3
}

export enum TileType {
    EMPTY = 'empty',
    PIT = 'pit',
    REPAIR = 'repair',
    OPTION = 'option',
    CONVEYOR = 'conveyor',
    EXPRESS_CONVEYOR = 'express',
    GEAR_CW = 'gear_cw',
    GEAR_CCW = 'gear_ccw',
    PUSHER = 'pusher'
}